added `channel.chat.clear` and `channel.chat.clear_user_messages`, moved `UserWhisperMessageHandler` to new file
