+++ TwitchLib.EventSub.Websockets/Core/EventArgs/Channel/ChannelChatClearArgs.cs
@@ -1,0 +1,7 @@
+﻿using TwitchLib.EventSub.Core.SubscriptionTypes.Channel;
+using TwitchLib.EventSub.Websockets.Core.Models;
+
+namespace TwitchLib.EventSub.Websockets.Core.EventArgs.Channel;
+
+public class ChannelChatClearArgs : TwitchLibEventSubEventArgs<EventSubNotification<ChannelChatClear>>
+{ }
