+++ TwitchLib.EventSub.Websockets/Core/EventArgs/Channel/ChannelChatClearUserMessagesArgs.cs
@@ -1,0 +1,7 @@
+﻿using TwitchLib.EventSub.Core.SubscriptionTypes.Channel;
+using TwitchLib.EventSub.Websockets.Core.Models;
+
+namespace TwitchLib.EventSub.Websockets.Core.EventArgs.Channel;
+
+public class ChannelChatClearUserMessagesArgs : TwitchLibEventSubEventArgs<EventSubNotification<ChannelChatClearUserMessage>>
+{ }
