+++ TwitchLib.EventSub.Websockets/Handler/Channel/Chat/ChannelChatClearHandler.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.Text.Json;
+using TwitchLib.EventSub.Core.SubscriptionTypes.Channel;
+using TwitchLib.EventSub.Websockets.Core.EventArgs;
+using TwitchLib.EventSub.Websockets.Core.EventArgs.Channel;
+using TwitchLib.EventSub.Websockets.Core.Handler;
+using TwitchLib.EventSub.Websockets.Core.Models;
+
+namespace TwitchLib.EventSub.Websockets.Handler.Channel.Chat;
+
+/// <summary>
+/// Handler for 'channel.chat.clear' notifications
+/// </summary>
+public class ChannelChatClearHandler : INotificationHandler
+{
+    /// <inheritdoc />
+    public string SubscriptionType => "channel.chat.clear";
+
+    /// <inheritdoc />
+    public void Handle(EventSubWebsocketClient client, string jsonString, JsonSerializerOptions serializerOptions)
+    {
+        try
+        {
+            var data = JsonSerializer.Deserialize<EventSubNotification<ChannelChatClear>>(jsonString.AsSpan(), serializerOptions);
+
+            if (data is null)
+                throw new InvalidOperationException("Parsed JSON cannot be null!");
+
+            client.RaiseEvent("ChannelChatClear", new ChannelChatClearArgs { Notification = data });
+        }
+        catch (Exception ex)
+        {
+            client.RaiseEvent("ErrorOccurred", new ErrorOccuredArgs { Exception = ex, Message = $"Error encountered while trying to handle {SubscriptionType} notification! Raw Json: {jsonString}" });
+        }
+    }
+}
