+++ TwitchLib.EventSub.Websockets/Handler/Channel/Chat/ChannelChatClearUserMessagesHandler.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.Text.Json;
+using TwitchLib.EventSub.Core.SubscriptionTypes.Channel;
+using TwitchLib.EventSub.Websockets.Core.EventArgs;
+using TwitchLib.EventSub.Websockets.Core.EventArgs.Channel;
+using TwitchLib.EventSub.Websockets.Core.Handler;
+using TwitchLib.EventSub.Websockets.Core.Models;
+
+namespace TwitchLib.EventSub.Websockets.Handler.Channel.Chat;
+
+/// <summary>
+/// Handler for 'channel.chat.clear_user_messages' notifications
+/// </summary>
+public class ChannelChatClearUserMessagesHandler : INotificationHandler
+{
+    /// <inheritdoc />
+    public string SubscriptionType => "channel.chat.clear_user_messages";
+
+    /// <inheritdoc />
+    public void Handle(EventSubWebsocketClient client, string jsonString, JsonSerializerOptions serializerOptions)
+    {
+        try
+        {
+            var data = JsonSerializer.Deserialize<EventSubNotification<ChannelChatClearUserMessage>>(jsonString.AsSpan(), serializerOptions);
+
+            if (data is null)
+                throw new InvalidOperationException("Parsed JSON cannot be null!");
+
+            client.RaiseEvent("ChannelChatClearUserMessages", new ChannelChatClearUserMessagesArgs { Notification = data });
+        }
+        catch (Exception ex)
+        {
+            client.RaiseEvent("ErrorOccurred", new ErrorOccuredArgs { Exception = ex, Message = $"Error encountered while trying to handle {SubscriptionType} notification! Raw Json: {jsonString}" });
+        }
+    }
+}
