+++ TwitchLib.EventSub.Websockets/Handler/User/UserUpdateHandler.cs
@@ -34,30 +34,4 @@ namespace TwitchLib.EventSub.Websockets.Handler.User
             }
         }
     } 
-    /// <summary>
-    /// Handler for 'user.whisper.message' notifications
-    /// </summary>
-    public class UserWhisperMessageHandler : INotificationHandler
-    {
-        /// <inheritdoc />
-        public string SubscriptionType => "user.whisper.message";
-
-        /// <inheritdoc />
-        public void Handle(EventSubWebsocketClient client, string jsonString, JsonSerializerOptions serializerOptions)
-        {
-            try
-            {
-                var data = JsonSerializer.Deserialize<EventSubNotification<UserWhisperMessage>>(jsonString.AsSpan(), serializerOptions);
-
-                if (data is null)
-                    throw new InvalidOperationException("Parsed JSON cannot be null!");
-
-                client.RaiseEvent("UserWhisperMessage", new UserWhisperMessageArgs { Notification = data });
-            }
-            catch (Exception ex)
-            {
-                client.RaiseEvent("ErrorOccurred", new ErrorOccuredArgs { Exception = ex, Message = $"Error encountered while trying to handle {SubscriptionType} notification! Raw Json: {jsonString}" });
-            }
-        }
-    }
 }
